Fixed bugs in remove_adjacent_letters()
There was a bug that would miss patters such as
_['6','7','6','7','7','8']_. It would return _['6','6','7','7','8']_
instead of the expected _['6','7','6','7','8']_. The bug was fixed.

If an empty list was passed into remove_adjacent_letters() the program
would crash due to an index error. An empty list could be passed in
some situations such as with the name Joe. The 'J' would be removed and
the two remaining letters are vowls which means they will also be removed
resulting in an empty list. An empty list now returns `['0', '0', '0']`.
The soundex for _Joe_ is _J000_.